requirements.txt
pywebview>=4.0.0
flask>=2.0.0
